Integrate downloadFeedFiles into subscription, with more checks.
Be very verbose (at least at this stage).
Include array of what days to check, which will be used to decide when
to download new podcasts.
Make getLatestEntryHelper/getLatestEntry generic instead of latest.
We want to be able to support downloading a podcast's backlog.
Next step is function to trigger a check of the podcast.